refactor!: remove deprecated GeneratedVaadinUpload
